[Feat] Connector (#66)
# Summary
The Connector intends to be a flexible point of interconnection and management for a user’s sockets. Through simple YAML definitions, users will be able to automatically create and run multiple sockets of different types or configurations.

The intent is that users don’t have to manually create sockets, but instead can define their intentions in a config file, and the connector will take care of everything that’s needed. It will create sockets, create tunnels, and set up the tunnel. As well as clean this up when no longer needed.

# Connector plugins

**Static sockets:** ability to define targets (DNS/IP and port numbers) in a config file. The connector will manage sockets for them.
**EC2 discovery:** connector will listen for events, or periodically scan the AWS ec2 API for instances with a mysocket tag. The connector will manage sockets for instances with the correct tag.
**Docker discovery:** connector will listen for events, or periodically scan the Docker API (Unix socket)  for containers with a mysocket label. The connector will manage sockets for containers with the correct tag.
**Network scanner:** connector will periodically scan local subnets for a list of predefined ports. The connector will manage sockets for IP and ports that match the config.
**Kubernetes k8s**: The connector will listen for events on the k8 API and find pods, The connector will manage sockets for instances with the correct tag.